compile warnings: signed vs unsigned and unused

+++ firmware/controllers/sensors/sensor_checker.cpp
@@ -172,7 +172,7 @@ void SensorChecker::onSlowCallback() {
 // only bother checking these if we have GPIO chips actually capable of reporting an error
 #if BOARD_EXT_GPIOCHIPS > 0 && EFI_PROD_CODE
 	// Check injectors
-	for (int i = 0; i < efi::size(enginePins.injectors); i++) {
+	for (size_t i = 0; i < efi::size(enginePins.injectors); i++) {
 		InjectorOutputPin& pin = enginePins.injectors[i];
 
 		// Skip not-configured pins
@@ -191,7 +191,7 @@ void SensorChecker::onSlowCallback() {
 	}
 
 	// Check ignition
-	for (int i = 0; i < efi::size(enginePins.injectors); i++) {
+	for (size_t i = 0; i < efi::size(enginePins.injectors); i++) {
 		IgnitionOutputPin& pin = enginePins.coils[i];
 
 		// Skip not-configured pins
